Added a class called 'Experiment'
The class models an object that allows the client to specify a set of pre learning processors (scaling, null-value resolution, feature selection), a set of machine learning methods, and a dataset. The client can then run experiments on every pair (or pipeline) of a pre learning processor and machine learning and obtain the results.